apps/api/requirements.txt
alembic==1.14.0
annotated-doc==0.0.4
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.24.0
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asyncpg==0.31.0
bcrypt==5.0.0
cffi==2.0.0
click==8.3.1
ecdsa==0.19.1
fastapi==0.135.1
greenlet==3.3.2
h11==0.16.0
idna==3.11
passlib==1.7.4
psycopg2-binary==2.9.11
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydantic-settings==2.0.1
email-validator==2.3.0
PyJWT==2.11.0
python-jose==3.5.0
python-multipart==0.0.22
redis==7.2.1
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.48
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
